requirements.txt
numpy
opencv-python
pillow
keras
matplotlib
tensorflow
